requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
requests>=2.27.0
pydantic>=1.8.2
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
openai>=1.0.0
python-jose[cryptography]
passlib[bcrypt]
aiosqlite 